pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Some LangGraph Examples"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph (>=0.6.4,<0.7.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.29,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "fastmcp>=2.11.2",
    "aiohttp>=3.12.15",
    "langchain-mcp-adapters>=0.1.9",
    "ipython>=9.4.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.12.8",
]
